==> experiments/7_evaluation.ipynb <==
import os
os.chdir("..")

import json
import math
import pandas as pd
import numpy as np
from tqdm import tqdm
import seaborn as sns
import matplotlib.pyplot as plt
from matplotlib.ticker import MaxNLocator
from sklearn.metrics import classification_report, accuracy_score, f1_score, recall_score, precision_score

tqdm.pandas() 

import warnings
warnings.filterwarnings('ignore')

# Experiments from paper

## Helper functions

def get_median(distr, start, end, step):
    x = np.arange(start, end, step)
    exp_f = np.exp(distr)
    alpha = exp_f.sum() * 0.5
    try:
        median_low = x[exp_f.cumsum() <= alpha][-1]
    except IndexError:
        median_low = start
    try:
        median_high = x[::-1][exp_f[::-1].cumsum() < alpha][-1]
    except IndexError:
        median_high = end
    median_avg = (median_low + median_high) / 2
    return median_avg

def get_thresholds_from_record(record, start=0, end=1, step=0.01):
    thresholds = dict()
    for claim in record:
        thresholds[claim] = {'threshold': get_median(record[claim]['distributions'][-1], start, end, step)}
    return thresholds

def keep_only_top_claim(scores_dict):
    new_dict = dict()
    
    for cl in CLASSES:
        rel_claims = [t for t in claims if t[:3] == cl]
        cl_dict = {t: scores_dict[t] for t in rel_claims}
        highest_t = max(cl_dict, key = cl_dict.get)
        for t in rel_claims:
            if t == highest_t:
                new_dict[t] = scores_dict[t]
            else:
                new_dict[t] = 0
    return new_dict

def get_multi_pred(row, column, claims_dict, thresholds=None, black_list=None, min_length=1):
    if thresholds is None:
        thresholds = {t: {'threshold': 0.0} for t in claims_dict}
    if black_list is None:
        black_list = []
    result = list()
    if isinstance(row[column], list):
        zsl_scores = {k:row[column][0][claims_dict[k]] for k in claims_dict}
    else:
        zsl_scores = {k:row[column][claims_dict[k]] for k in claims_dict}
    for c in CLASSES:
        sub_pred = {k: zsl_scores[k] for k in zsl_scores if k[:3] == c and not k in black_list}
        if len([k for k in sub_pred if sub_pred[k] > thresholds[k]['threshold']]) >= min_length:
            result.append(1)
        else:
            result.append(0)
    return result


def get_binary_pred(row, claim_class, column, claims_dict, thresholds=None, black_list=None, min_length=1):
    if thresholds is None:
        thresholds = {t: {'threshold': 0.0} for t in claims_dict}
    if black_list is None:
        black_list = []
    result = list()
    if isinstance(row[column], list):
        zsl_scores = {k:row[column][0][claims_dict[k]] for k in claims_dict}
    else:
        zsl_scores = {k:row[column][claims_dict[k]] for k in claims_dict}

    sub_pred = {k: zsl_scores[k] for k in zsl_scores if k[:3] == claim_class and not k in black_list}
    if len([k for k in sub_pred if sub_pred[k] > thresholds[k]['threshold']]) >= min_length:
        return 1
    else:
        return 0

## Climate Change Contrarianism detection

test_df = pd.read_pickle('./data/climate_change/testing.pkl')

CLASSES = ['1_1', '1_2', '1_3', '1_4', '1_6', '1_7']

with open("./data/climate_change/claims.json") as file:
    claims = json.load(file)

claims_descr = claims["class_descr"]
del claims["class_descr"]

inverse_claims = {claims[k]: k for k in claims}

### Baseline 1: Fine-tuned BERT

print(classification_report(test_df['multi_annot'].to_list(), test_df['BERT'].to_list(), target_names=CLASSES))

### Baseline 2: SBERT cosine similarity with threshold-tuning

with open('./data/bisection_records/CCC_SBERT.json') as file:
    sbert_record = json.load(file)

sbert_thresholds = get_thresholds_from_record(sbert_record, start=-1, end=1)

test_df['pred_multi'] = test_df.apply(lambda x: get_multi_pred(x, 'sbert_cosine', claims, thresholds=sbert_thresholds), axis=1)

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

datapoints = list()
for k in sbert_record:
    datapoints += sbert_record[k]['texts']
len(set(datapoints))

### Baseline 3: BART MNLI model with unique threshold (Zero-Shot approach)

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

test_df['pred_multi'] = test_df.apply(lambda x: get_multi_pred(x, 'FSL_BART', claims, thresholds=artificial_thresholds), axis=1)

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

### Baseline 4: BART MNLI model with Temperature Scaling

sample_sizes = [5, 10, 20, 40, 80, 160]

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

for samp_size in sample_sizes:
    test_df['pred_multi'] = test_df.apply(lambda x: get_multi_pred(x, 'Temp_Scaling_BART_' + str(samp_size), claims, thresholds=artificial_thresholds), axis=1)
    print(samp_size, ': %.3f' % f1_score(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), average="micro"))

### Baseline 5: Llama 2 70b model with prompting

def get_llama_pred(row, column, claims_dict, black_list=None, min_length=1):
    if black_list is None:
        black_list = []
    result = list()
    zsl_scores = {k:row[column][claims_dict[k]] for k in claims_dict}
    for c in CLASSES:
        sub_pred = {k: zsl_scores[k] for k in zsl_scores if k[:3] == c and not k in black_list}
        if len([k for k in sub_pred if sub_pred[k].lower().strip() == "yes"]) >= min_length:
            result.append(1)
        else:
            result.append(0)
    return result

test_df['pred_multi'] = test_df.apply(lambda x: get_llama_pred(x, 'Llama-2', claims), axis=1)

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

### Ours 1: Few-shots NLI approach using BART MNLI

with open('./data/bisection_records/CCC_BART.json') as file:
    bart_record = json.load(file)

bart_thresholds = get_thresholds_from_record(bart_record, start=0, end=1)

test_df['pred_multi'] = test_df.apply(lambda x: get_multi_pred(x, 'FSL_BART', claims, thresholds=bart_thresholds), axis=1)

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

datapoints = list()
for k in bart_record:
    if k in ["1_4_0_1"]:
        continue
    datapoints += bart_record[k]['texts']
len(set(datapoints))

## Topic and stance classification

test_df = pd.read_pickle('./data/topic_stance/testing.pkl')

CLASSES_TOPIC = ['1', '2', '3', '4', '5']
CLASSES_STANCE = ['1A', '1F', '1N', '2A', '2F', '2N', '3A', '3F', '3N', '4A', '4F', '4N', '5A', '5F', '5N']

with open("./data/topic_stance/claims_topic.json") as file:
    claims_topic = json.load(file)
    
with open("./data/topic_stance/claims_stance.json") as file:
    claims_stance = json.load(file)
    
claims = {**claims_topic, **claims_stance}

claims_descr = claims["class_descr"]
del claims["class_descr"]

inverse_claims = {claims[k]: k for k in claims}

def custom_transformation(x, threshold):
    if x <= threshold:
        return x * (0.5 / threshold)
    else:
        return 0.5 + (x - threshold) * (0.5 / (1 - threshold))
    
    
def get_normed_scores(zsl_dict, thresholds, inverse_claims):
    new_results = dict()
    for trait in zsl_dict:
        if not trait in inverse_claims:
            continue
        new_results[trait] = custom_transformation(zsl_dict[trait], thresholds[inverse_claims[trait]]['threshold'])
    return new_results


def get_avg_normed_topic_scores(normed_zsl_dict, classes_list, inverse_claims):
    new_results = dict()
    for c in classes_list:
        temp = [normed_zsl_dict[trait] for trait in normed_zsl_dict if inverse_claims[trait].split("_")[0] == c]
        if len(temp) > 0:
            new_results[c] = sum(temp) / len(temp)
    return new_results
    

def get_topic_pred_multiclass(avg_normed_ZSL_scores):
    return max(avg_normed_ZSL_scores, key=avg_normed_ZSL_scores.get)

def get_stance_pred(normed_zsl_dict, topic, traits):
    traits_anti = [traits[t] for t in traits if t[:2] == (topic + "A")]
    detected_anti = [normed_zsl_dict[t] for t in traits_anti if normed_zsl_dict[t] > 0.5]
    traits_favor = [traits[t] for t in traits if t[:2] == (topic + "F")]
    detected_favor = [normed_zsl_dict[t] for t in traits_favor if normed_zsl_dict[t] > 0.5]
    if len(detected_anti) > len(detected_favor):
        return topic + "A"
    elif len(detected_favor) > len(detected_anti):
        return topic + "F"
    if len(detected_favor) == 0:
        return topic + "N"
    avg_scores_anti = sum(detected_anti) / len(detected_anti)
    avg_scores_favour = sum(detected_favor) / len(detected_favor)
    if avg_scores_anti > (1.05 * avg_scores_favour):
        return topic + "A"
    elif avg_scores_favour > (1.05 * avg_scores_anti):
        return topic + "F"
    return topic + "N"
    return max(normed_zsl_dict, key=normed_zsl_dict.get)

### Baseline 1: Fine-tuned BERT

print(classification_report(test_df['topic_stance_annot'].to_list(), test_df['BERT_topic_stance'].to_list(), target_names=CLASSES_STANCE))

### Baseline 2: SBERT cosine similarity with threshold-tuning

with open('./data/bisection_records/TS_SBERT_topic.json') as file:
    sbert_record_topic = json.load(file)
    
with open('./data/bisection_records/TS_SBERT_stance.json') as file:
    sbert_record_stance = json.load(file)

sbert_topic_thresholds = get_thresholds_from_record(sbert_record_topic, start=-1, end=1)
sbert_stance_thresholds = get_thresholds_from_record(sbert_record_stance, start=-1, end=1)

test_df['normed_sbert_topic'] = test_df["sbert_cosine_topic"].apply(lambda x: get_normed_scores(x, sbert_topic_thresholds, inverse_claims))
test_df['normed_sbert_stance'] = test_df["sbert_cosine_stance"].apply(lambda x: get_normed_scores(x, sbert_stance_thresholds, inverse_claims))
test_df['avg_normed_sbert_topic'] = test_df["normed_sbert_topic"].apply(lambda x: get_avg_normed_topic_scores(x, CLASSES_TOPIC, inverse_claims))
test_df['pred_topic'] = test_df['avg_normed_sbert_topic'].apply(get_topic_pred_multiclass)
test_df['pred_stance'] = test_df.apply(lambda x: get_stance_pred(x['normed_sbert_stance'], x['pred_topic'], claims_stance), axis=1)

print(classification_report(test_df['topic_stance_annot'].to_list(), test_df['pred_stance'].to_list(), target_names=CLASSES_STANCE))

datapoints = list()
for k in sbert_record_topic:
    datapoints += sbert_record_topic[k]['texts']
    
for k in sbert_record_stance:
    datapoints += sbert_record_stance[k]['texts']
    
len(set(datapoints))

### Baseline 3: BART MNLI model with unique threshold (Zero-Shot approach)

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

test_df['normed_BART_topic'] = test_df["FSL_BART_topic"].apply(lambda x: get_normed_scores(x, artificial_thresholds, inverse_claims))
test_df['normed_BART_stance'] = test_df["FSL_BART_stance"].apply(lambda x: get_normed_scores(x, artificial_thresholds, inverse_claims))
test_df['avg_normed_BART_topic'] = test_df["normed_BART_topic"].apply(lambda x: get_avg_normed_topic_scores(x, CLASSES_TOPIC, inverse_claims))
test_df['pred_topic'] = test_df['avg_normed_BART_topic'].apply(get_topic_pred_multiclass)
test_df['pred_stance'] = test_df.apply(lambda x: get_stance_pred(x['normed_BART_stance'], x['pred_topic'], claims_stance), axis=1)

print(classification_report(test_df['topic_stance_annot'].to_list(), test_df['pred_stance'].to_list(), target_names=CLASSES_STANCE))

### Baseline 4: BART MNLI model with Temperature Scaling

sample_sizes = [5, 10, 20, 40, 80, 160]

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

test_df.columns

for samp_size in sample_sizes:
    test_df['normed_TempScal_topic'] = test_df["Temp_Scaling_BART_" + str(samp_size) + "_topic"].apply(lambda x: get_normed_scores(x, artificial_thresholds, inverse_claims))
    test_df['normed_TempScal_stance'] = test_df["Temp_Scaling_BART_" + str(samp_size) + "_stance"].apply(lambda x: get_normed_scores(x, artificial_thresholds, inverse_claims))
    test_df['avg_normed_TempScal_topic'] = test_df["normed_TempScal_topic"].apply(lambda x: get_avg_normed_topic_scores(x, CLASSES_TOPIC, inverse_claims))
    test_df['pred_topic'] = test_df['avg_normed_TempScal_topic'].apply(get_topic_pred_multiclass)
    test_df['pred_stance'] = test_df.apply(lambda x: get_stance_pred(x['normed_TempScal_stance'], x['pred_topic'], claims_stance), axis=1)
    print(samp_size, ': %.3f' % f1_score(test_df['topic_stance_annot'].to_list(), test_df['pred_stance'].to_list(), average="micro"))

### Baseline 5: Llama 2 70b model with prompting

test_df.columns

target2idx = {
    "Atheism": "1",
    "Climate Change is a Real Concern": "2",
    "Feminist Movement": "3",
    "Hillary Clinton": "4",
    "Legalization of Abortion": "5"
}

def get_llama_pred(stance_preds, stance_annots, topic_annot):
    result = list()
    
    for topic in target2idx:
        if topic in stance_preds and topic == topic_annot: 
            if stance_preds[topic].lower() == "anti":
                result += [1, 0, 0]
            elif stance_preds[topic].lower() == "pro":
                result += [0, 1, 0]
            else:
                result += [0, 0, 1]
        else:
            result += [0, 0, 0]
    
    sub_results = list()
    
    for p,a in zip(result, stance_annots):
        if a == 0:
            sub_results.append(0)
        else:
            sub_results.append(p)
            
    return sub_results

test_df['topic_stance_annot_multi'] = test_df['topic_stance_annot'].apply(lambda x: [1 if k == x else 0 for k in CLASSES_STANCE])

test_df['pred_stance'] = test_df.apply(lambda x: get_llama_pred(x["Llama-2_STANCE"], x["topic_stance_annot_multi"], x["Target"]), axis=1)

print(classification_report(test_df['topic_stance_annot_multi'].to_list(), test_df['pred_stance'].to_list(), target_names=CLASSES_STANCE))

### Ours 1: Few-shots NLI approach using BART MNLI

with open('./data/bisection_records/TS_BART_topic.json') as file:
    bart_record_topic = json.load(file)
    
with open('./data/bisection_records/TS_BART_stance.json') as file:
    bart_record_stance = json.load(file)

bart_topic_thresholds = get_thresholds_from_record(bart_record_topic, start=0, end=1)
bart_stance_thresholds = get_thresholds_from_record(bart_record_stance, start=0, end=1)

bart_topic_thresholds

test_df['normed_BART_topic'] = test_df["FSL_BART_topic"].apply(lambda x: get_normed_scores(x, bart_topic_thresholds, inverse_claims))
test_df['normed_BART_stance'] = test_df["FSL_BART_stance"].apply(lambda x: get_normed_scores(x, bart_stance_thresholds, inverse_claims))
test_df['avg_normed_BART_topic'] = test_df["normed_BART_topic"].apply(lambda x: get_avg_normed_topic_scores(x, CLASSES_TOPIC, inverse_claims))
test_df['pred_topic'] = test_df['avg_normed_BART_topic'].apply(get_topic_pred_multiclass)
test_df['pred_stance'] = test_df.apply(lambda x: get_stance_pred(x['normed_BART_stance'], x['pred_topic'], claims_stance), axis=1)

print(classification_report(test_df['topic_stance_annot'].to_list(), test_df['pred_stance'].to_list(), target_names=CLASSES_STANCE))

datapoints = list()
for k in bart_record:
    if k in ["1_4_0_1"]:
        continue
    datapoints += bart_record[k]['texts']
len(set(datapoints))

## Depressive symptoms detection

test_df = pd.read_pickle('./data/depression/testing.pkl')

test_df = test_df.drop_duplicates(subset=["Sentence"])

CLASSES = [
    '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', 
    '17', '18', '19', '20', '21'
]

with open("./data/depression/claims.json") as file:
    claims = json.load(file)

claims_descr = claims["class_descr"]
del claims["class_descr"]

inverse_claims = {claims[k]: k for k in claims}

test_df["multi_annot"] = test_df.apply(lambda x: [x[C + "_annot"] for C in CLASSES], axis=1)

def get_detected_claims(zsl_dict, thresholds, claims):
    detected = list()
    for claim_idx in claims:
        if zsl_dict[claims[claim_idx]] >= thresholds[claim_idx]['threshold']:
            detected.append(claim_idx)
    return detected 


def get_pred(detected, classes):
    preds = list()
    for group in classes:
        rel = [d for d in detected if d.split("_")[0] == group]
        if len(rel) > 0:
            preds.append(1)
        else:
            preds.append(0)
    return preds

### Baseline 1: Fine-tuned BERT

print(classification_report(test_df['multi_annot'].to_list(), test_df['BERT_pred'].to_list(), target_names=CLASSES))

### Baseline 2: SBERT cosine similarity with threshold-tuning

with open('./data/bisection_records/D_SBERT.json') as file:
    sbert_record = json.load(file)

sbert_thresholds = get_thresholds_from_record(sbert_record, start=-1, end=1)

test_df['detected'] = test_df['sbert_cosine'].apply(lambda x: get_detected_claims(x, sbert_thresholds, claims))
test_df['pred_multi'] = test_df['detected'].apply(lambda x: get_pred(x, CLASSES))

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

datapoints = list()
for k in sbert_record:
    datapoints += sbert_record[k]['texts']
len(set(datapoints))

### Baseline 3: BART MNLI model with unique threshold (Zero-Shot approach)

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

test_df['detected'] = test_df['FSL_BART'].apply(lambda x: get_detected_claims(x, artificial_thresholds, claims))
test_df['pred_multi'] = test_df['detected'].apply(lambda x: get_pred(x, CLASSES))

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

### Baseline 4: BART MNLI model with Temperature Scaling

sample_sizes = [5, 10, 20, 40, 80, 160]

artificial_thresholds = {k: {'threshold': 0.5} for k in claims}

for samp_size in sample_sizes:
    test_df['detected'] = test_df["Temp_Scaling_BART_" + str(samp_size)].apply(lambda x: get_detected_claims(x, artificial_thresholds, claims))
    test_df['pred_multi'] = test_df['detected'].apply(lambda x: get_pred(x, CLASSES))
    print(samp_size, ': %.3f' % f1_score(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), average="micro"))

### Baseline 5: Llama 2 (70b model) with prompting

def get_detected_claims_LLAMA(zsl_dict, claims):
    detected = list()
    for claim_idx in claims:
        if zsl_dict[claims[claim_idx]].lower().strip() == "yes":
            detected.append(claim_idx)
    return detected 

test_df['detected'] = test_df['Llama-2'].apply(lambda x: get_detected_claims_LLAMA(x, claims))
test_df['pred_multi'] = test_df['detected'].apply(lambda x: get_pred(x, CLASSES))

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

### Ours 1: Few-shots NLI approach using BART MNLI

with open('./data/bisection_records/D_BART.json') as file:
    bart_record = json.load(file)

bart_thresholds = get_thresholds_from_record(bart_record, start=0, end=1)

test_df['detected'] = test_df['FSL_BART'].apply(lambda x: get_detected_claims(x, bart_thresholds, claims))
test_df['pred_multi'] = test_df['detected'].apply(lambda x: get_pred(x, CLASSES))

print(classification_report(test_df['multi_annot'].to_list(), test_df['pred_multi'].to_list(), target_names=CLASSES))

datapoints = list()
for k in bart_record:
    if k in ["1_4_0_1"]:
        continue
    datapoints += bart_record[k]['texts']
len(set(datapoints))
